api/requirements.txt
fastapi
pillow
requests
tensorflow
uvicorn[standard]
